+++ packages/@biomejs/backend-jsonrpc/src/workspace.ts
@@ -1164,6 +1164,10 @@ export interface Nursery {
 	 */
 	noDuplicateAtImportRules?: RuleConfiguration_for_Null;
 	/**
+	 * Disallow duplicate custom properties within declaration blocks.
+	 */
+	noDuplicateCustomProperties?: RuleConfiguration_for_Null;
+	/**
 	 * Disallow duplicate conditions in if-else-if chains
 	 */
 	noDuplicateElseIf?: RuleConfiguration_for_Null;
@@ -2748,6 +2752,7 @@ export type Category =
 	| "lint/nursery/noConsole"
 	| "lint/nursery/noDoneCallback"
 	| "lint/nursery/noDuplicateAtImportRules"
+	| "lint/nursery/noDuplicateCustomProperties"
 	| "lint/nursery/noDuplicateElseIf"
 	| "lint/nursery/noDuplicateFontNames"
 	| "lint/nursery/noDuplicateSelectorsKeyframeBlock"